requirements.txt
Flask 
Flask-SQLAlchemy
SQLAlchemy
openai
Flask-Migrate
pytest
python-dotenv